Moved findTransactionsByAccountAndTagIds service function from tag service file to transaction service file.
