+++ src/services/transaction.service.ts
@@ -1,6 +1,7 @@
 import {Transaction} from '../entities/transaction';
 import {Account} from '../entities/account';
 import {dataSource} from '../utilities/data-source';
+import {Tag} from '../entities/tag';
 
 export const createTransaction = async (transaction: Transaction) => {
   if (transaction.account) {
@@ -37,3 +38,16 @@ export const findAllTransactionsByAccount = async (account: Account) =>
     order: {id: 'DESC'},
     relations: {tags: true, account: true},
   });
+
+export const findTransactionsByAccountAndTagIds = async (
+  accountId: Account['id'],
+  tagId: Tag['id'],
+) =>
+  await dataSource
+    .getRepository(Transaction)
+    .createQueryBuilder('transaction')
+    .innerJoin('transaction.tags', 'tag')
+    .where('transaction.accountId = :accountId', {accountId})
+    .andWhere('tag.id = tagId', {tagId})
+    .orderBy('transaction.id', 'DESC')
+    .getMany();
